[DEV] Change the UI according to the connected wifi network
In the About fragment, if the currently connected wifi network
is not the one of the venue, show a view to connect to it.
Once the user has touched this view, show a loading wheel until
the connection has succeed.
WHen the current wifi network is the one of the venue, hide both
the view and the loading wheel.