[DEV] Reworked notifications for starred sessions
Show the room on the notifications
Allow multiple notifications if there are several starred
sessions that begins at the same time.
Remove the alarm when a session is unstarred.
When the data source is updated, recreate all alarms
in case the schedule has changed.
Fixed a bug where we cannot fetch the session or the slot
information because the AgendaRepository was not ready.